Extract PID logic from TurningRateClosedLoopProcessor
Also documentation!
